Expand exception handling (#42)
Expand Env exception handling by adding the ExceptionHandler
interface and ExceptionHandler field to Env. When this field is set
in Env, exceptions encountered and run through handleException are
processed using the provided handler, and the error it creates is
returned. This enables custom error handling / exception conversion.

In addition to providing custom handling, this patch includes three
handlers available for use:

- DescribeExceptionHandler provides the previous behavior, printing
  with describeException.
- ThrowableToStringExceptionHandler calls the "toString" method of
  the exception, and returns an error with that message.
- ThrowableErrorExceptionHandler retrieves the values in the
  exception, storing them in the new ThrowableError error, and
  returning that value. This conversion includes the stack trace,
  converting those values into a slice of the new StackTraceElement
  struct. This is recursive to convert any underlying cause.

By default, the existing behavior is preserved, and adding one of the
new handlers (or a custom one) is opt-in.